# api/production/openapi.yaml
openapi: 3.0.3
info:
  title: E.Y.T Production API
  version: 0.1.0
  description: Production lifecycle API for E.Y.T ERP.
servers:
  - url: /api
paths:
  /production/orders:
    post:
      summary: Create production order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionOrderInput'
      responses:
        '201':
          description: Created
  /production/orders/{orderNo}:
    get:
      summary: Get production order
      parameters:
        - $ref: '#/components/parameters/OrderNo'
      responses:
        '200':
          description: Production order
        '404':
          description: Not found
...
components:
  parameters:
    OrderNo:
      name: orderNo
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    ProductionOrderInput:
      type: object
      required: [orderNo, productCode, productName, targetQty, orderDate]
      properties:
        orderNo: { type: string }
        productCode: { type: string }
        productName: { type: string }
        targetQty: { type: number }
        orderDate: { type: string, format: date }
        plannedStart: { type: string, format: date }
        plannedEnd: { type: string, format: date }
        customerId: { type: integer, nullable: true }
...